feat: Refactor account repository
