+++ src/core/application/use-cases/accounts/fetch-all-account-use-case.ts
@@ -3,7 +3,7 @@ import { PaginationEntity } from '@/core/domain/entities/pagination-entity'
 import { AccountMapper } from '../../mappers/account-mapper'
 import { AccountEntity } from '@/core/domain/entities/account-entity'
 import { AccountResponseDto } from '../../dto/account-dto'
-import { FetchAllAccountsRepository } from '@/core/domain/repositories/accounts/fetch-all-accounts-repository'
+import { AccountRepository } from '@/core/domain/repositories/account-repository'
 import { inject, injectable } from 'inversify'
 import { LogOperation } from '../../decorators/log-operation'
 
@@ -19,8 +19,8 @@ export interface FetchAllAccounts {
 @injectable()
 export class FetchAllAccountsUseCase implements FetchAllAccounts {
   constructor(
-    @inject(FetchAllAccountsRepository)
-    private readonly fetchAllAccountsRepository: FetchAllAccountsRepository
+    @inject(AccountRepository)
+    private readonly accountRepository: AccountRepository
   ) {}
 
   @LogOperation({ layer: 'application' })
@@ -31,7 +31,7 @@ export class FetchAllAccountsUseCase implements FetchAllAccounts {
     page: number
   ): Promise<PaginationDto<AccountResponseDto>> {
     const accountsResult: PaginationEntity<AccountEntity> =
-      await this.fetchAllAccountsRepository.fetchAll(
+      await this.accountRepository.fetchAll(
         organizationId,
         ledgerId,
         page,
