--- src/core/domain/repositories/accounts/create-accounts-repository.ts
@@ -1,9 +1,0 @@
-import { AccountEntity } from '../../entities/account-entity'
-
-export abstract class CreateAccountsRepository {
-  abstract create: (
-    organizationId: string,
-    ledgerId: string,
-    account: AccountEntity
-  ) => Promise<AccountEntity>
-}
