--- src/core/domain/repositories/accounts/delete-accounts-repository.ts
@@ -1,7 +1,0 @@
-export abstract class DeleteAccountsRepository {
-  abstract delete: (
-    organizationId: string,
-    ledgerId: string,
-    accountId: string
-  ) => Promise<void>
-}
