--- src/core/domain/repositories/accounts/fetch-account-by-id-repository.ts
@@ -1,9 +1,0 @@
-import { AccountEntity } from '../../entities/account-entity'
-
-export abstract class FetchAccountByIdRepository {
-  abstract fetchById: (
-    organizationId: string,
-    ledgerId: string,
-    accountId: string
-  ) => Promise<AccountEntity>
-}
