--- src/core/domain/repositories/accounts/fetch-all-accounts-repository.ts
@@ -1,11 +1,0 @@
-import { AccountEntity } from '../../entities/account-entity'
-import { PaginationEntity } from '../../entities/pagination-entity'
-
-export abstract class FetchAllAccountsRepository {
-  abstract fetchAll: (
-    organizationId: string,
-    ledgerId: string,
-    limit: number,
-    page: number
-  ) => Promise<PaginationEntity<AccountEntity>>
-}
