+++ src/core/infrastructure/container-registry/midaz/midaz-module.ts
@@ -1,5 +1,5 @@
+import { AccountRepository } from '@/core/domain/repositories/account-repository'
 import { Container, ContainerModule } from '../../utils/di/container'
-import { MidazAccountModule } from './account-module'
 import { MidazAssetModule } from './asset-module'
 import { MidazLedgerModule } from './ledger-module'
 import { MidazOrganizationModule } from './organization-module'
@@ -7,13 +7,16 @@ import { MidazPortfolioModule } from './portfolio-module'
 import { MidazSegmentModule } from './segment-module'
 import { MidazTransactionModule } from './transaction-module'
 import { BalanceRepository } from '@/core/domain/repositories/balance-repository'
-import { MidazBalanceRepository } from '@/core/infrastructure/midaz/midaz-balance-repository'
+import { MidazBalanceRepository } from '@/core/infrastructure/midaz/repositories/midaz-balance-repository'
+import { MidazAccountRepository } from '@/core/infrastructure/midaz/repositories/midaz-account-repository'
 
 export const MidazModule = new ContainerModule((container: Container) => {
   container.load(MidazOrganizationModule)
   container.load(MidazLedgerModule)
   container.load(MidazPortfolioModule)
-  container.load(MidazAccountModule)
+  container
+    .bind<AccountRepository>(AccountRepository)
+    .to(MidazAccountRepository)
   container.load(MidazAssetModule)
   container.load(MidazSegmentModule)
   container.load(MidazTransactionModule)
